Fixed 230,247,99,302,261

+++ src/src/BioSounds/Controller/Administration/SiteCollectionController.php
@@ -38,14 +38,31 @@ class SiteCollectionController extends BaseController
      */
     public function save(): string
     {
+        $result = [];
         $siteCollectionProvider = new SiteCollection();
-        foreach ($_POST['d'] as $row) {
-            if (isset($row['c'])) {
-                $siteCollectionProvider->delete($row['c'], $_POST['site_id']);
-                if ($row['b'] == 'true') {
-                    $siteCollectionProvider->insert($row['c'], $_POST['site_id']);
+        $collection = explode(',', $siteCollectionProvider->isValid($_POST['site'], $_POST['site_id']));
+        if (isset($_POST['d'])) {
+            foreach ($_POST['d'] as $row) {
+                if (isset($row['c'])) {
+                    $siteCollectionProvider->delete($row['c'], $_POST['site_id']);
+                    if ($row['b'] == 'true') {
+                        if (in_array($row['c'], $collection)) {
+                            if (!in_array($row['p'], $result)) {
+                                $result[] = $row['p'];
+                            }
+                            continue;
+                        }
+                        $siteCollectionProvider->insert($row['c'], $_POST['site_id']);
+                    }
                 }
             }
+        }
+        if (count($result) > 0) {
+            return json_encode([
+                'isValid' => 1,
+                'result' => $result,
+                'message' => 'Site name already exists in the project.',
+            ]);
         }
         return json_encode([
             'errorCode' => 0,
